Updated urls to reflect the new templates and paths. Updated urltests to match. Wrote tests for the Login view class's get and post methods. The tests for the get method ensure that the correct template is loaded, that the user is logged out, and that other users are not logged out. The tests for the post method ensure that the correct template is retrieved depending on if the user is logged in or not, that the user is correctly logged in or not, and that invalid/weird input doesn't break it. We also test to see if multiple users can be logged in at once.
I also wrote the get and post methods. The get method simply logs out the current user using the prebuilt logout method, and returns us to the login page. The post method retrieves the entered information from the correct fields, and uses the prebuilt authenticate method to verify them. If the authenticate method verifies them, the user is logged in with the prebuilt login method and taken to the home page. Otherwise, the login page is reloaded.