Big update.
Wrote unit tests for the get and post methods on the user management page. The get tests mainly ensure that we are retrieving the correct user information from the database, and then displaying the correct information on the page itself. The post tests ensure that a new account is created, with user input being stored correctly. It also ensures that the user is added to the manager group and that the new user is display. Tests also ensure that the system works with multiple users and a variety of odd inputs.

The get method for the user page is relatively simple, it retrieves objects from the database and stores them based on group names. TAs are stored in a TA object, etc. These are then sent to the template, where they are displayed on the webpage.

The post method is slightly more complex. It retrieves the entered username, email, name and password from the form. It then searches the user database for a user with an identical username. If it finds one, no account is created. On the other hand, if it is unable to find one, then a new account is created, with the information from the form used to fill in fields. The new user is then added to the manager group and saved. Then, the accountmanagement page is reloaded so that the new user is displayed.

User management has had python code added to it to display current users. The code loops through the list of TAs, Instructors and administrators sent to the template. Other fields were added such as password and name to create a complete user object. I also altered the name attributes of some elements so that they are compatible with the get and post methods.

Course management template has been updated with early python code that indicates how the page will work. Forms were also added to handle adding sections and courses.