Another large and painful commit. This one contains unit tests for the get and post methods of the View class for the course management page. The functional implementations are also present. Changes were also made to the index.html document so that the links function correctly. CourseManagement.html was also modified, mainly to facilitate display existing courses/sections and creating new ones. Details can be seen below and are organized by section. I also created a models class so that I could test that this would all work, although it is no present within the commit. I can add/show it if anyone wants.
Get Unit Tests:
These tests mostly ensure that correct data is "loaded" as in present in the received response from the client call at the expected location. These test also ensure that the data is displayed, on the template. Simple cases are tested first, such as loading and displaying a course with a single user, section and section user. We then move into progressively more complex cases, testing multiple users, courses section users, sections and combinations of the above. Odd cases such as a course lacking users are also tested, although not as extensively.

Post Unit Tests:
These tests follow a similar pattern as the Get Unit Tests, as in they begin with relatively simple cases, such as creating a new course with no existing courses, and creating a new section for that course. We then enter more complex cases, such as creating new courses while multiple courses and sections other exist. We also test odd cases, such one or both inputs being missing, or an attempt to create a duplicate course.

Get Method:
This one is actually really simple, we just grab every Course and Section object, and throw them over to the template to figure out.

Post Method:
This one makes up for the lack of complexity in the Get Method unfortunately. We start out by checking if the "coursename" and "coursedescription" keys are present in the request. If they are, we then validate them, making sure neither one is empty. We then retrieve the desired course name and description from the request. We then utilize these to create a new course and then set it's users to an empty list.

If any of the above conditionals fail and the key "course" is present in the request, then we are creating a new section. We start by retrieving the course name string from the request, which is associated with the "course" key. We then use this name to find the actual object. Then, to determine the name of this new section, we retrieve the number of sections already associated with this course, add 1 to it, and turn it into a string. For example, a course with 1 section already associated with it would return 2 from this. Then, a loop is set up, while the counter+length of number of sections string is less than 3, 0 is added to a name string. Finally, after this, the string representing the number of sections+1 is added to the name string, giving us our final name. NOTE: This system does have a limit, it can only handle 999 sections at most, probably fine though. This newly created name is then used to create a section and assign it to the target course.

Regardless of whatever we do, we call a redirect to the /coursemanagement/ url to display whatever changes we made.

Index.html:
Relatively simple, just replaced the .html pages in the anchor tags href attributes with the urls in the url file. For whatever reason Django hates .html pages.

CourseManagement.html:
A surprising amount of work actually goes on here. This template receives a list of all Sections and Courses in the database. This template starts off with a for loop, going through every single Course received. Inside this loop, we display the name and description attribute of the current Course object. Following this, we run into a second loop, in which we go through every user stored in the current Course object's user field. These users are then displayed. Then, after this second loop, we hit a third loop, which goes through every Section object received. This loop immediately resets if the current Section object does not have the current Course object from the outer loop as its Course attribute. After this loop displays the name of each section attribute. Within this loop, we then have a fourth loop, which goes through the name of each user associated with a valid section, using the users attribute.

After these loops, we hit a small form. This form has a single input field, that is invisible. It's value is set to the name of the current Course object. The name attribute of this input is "course". Then, we have a submit button that calls the post method of the course management view class. The value of this button is "Create Section". Since the form is within the outer while loop, it will be populated within every course, so every course will have this tiny form that passes the name of the course to the post method, which will then create and empty section associated with the correct course.

Finally, after the main loop, we encounter a larger form. This one is more familiar, with inputs for a course name and description. The values of these fields are passed to the post method of the course creation View class, and used to create a course.