This one probably took a few years off my life to be honest. I reworked the get/post methods for user manangement and their unit tests to work with the new models.py class. I also wrote the unit tests for the get and post methods for the profile class, as well as the methods themselves. I also updated profile template so that it would work with my get/post methods. I also figured out how to make the CSS file work, after a lot of googling and finally dumb luck. Organized by section below if more details are desired.
Starting out with the rework of get/post methods for user management and their unit tests. I did have to make one small change to the models.py file, changing desc into description, and then propagating that change throughout the declared methods. Other than that, it was really just updating test methods to use the new Profile class.

Moving onto the Profile class, and its get and post method. The get method is relatively simple, we start out by assumming that the user is logged in. We use that to retrieve the logged in users id, with the session. Then we can find the current User object, and with that we can find the current Profile object. Using this, we can retrieve the desired information (name, email, altemail, address, etc) and store it in variables. Then we can map those variables to strings and send them off to the profile template. Simple!

The get method is also relatively simple, it just retrieves data from the profile template, utilizes the same method as above to find the current User and Profile object. Then, we check all retrieved, input, if the retrieved string is not empty, we update the User or Profile object accordingly. The only exception is the Username. We do an extra check to make sure an entered username does not conflict with an already established account. After this, we return a redirect to the page to display our new changes.

The unit tests for this class are not as extensive as some of the other ones I wrote. They mainly check for basic funtionality, and ensure that empty input, as well as normal input is processed as expected, and reflected on the template.

The changes to the profile template are fairly simple, I just added a few extra fields, made a form, and set up Django to print out the recieved information.

I also got the CSS to work. Apparently, we were meant to place it inside of a folder called static, inside our project directory, and then add {%load static%} at the top of each HTML document. Then, inside the href tag, we were meant to place {% static 'styles/style.css' %}. I can happily confirm that the CSS does now work on all pages.

Apologies for any typos, as I am half-dead.