General Changes (That got applied to whole site): Color scheme change from blues to black and blues. h1 header is now centered along with the nav menu.
In coursemanagement I added a 'delete course' button which can be seen alongside the course name.

In usermanagement I added a 'assign course' to Instructor and TA, and a 'assign section' to TA.
Neither of these can be seen currently so I don't know exactly what the button will look like until they are implemented. In theory, they will be alongside the name and role.