Refactored getMethod, and introduced a few helper methods, as well as a new Model class.
The getMethod now retrieves the current User object by filtering the list of existing users by the current user id in the session. This User object is then passed into the load helper method, which returns lists of User objects, Profile Objects, and two Button Lists to pass into the template. The User objects are then filtered into TA, Instructor and Admin Lists, using the determineRole method.

This method then returns a render of the usermanagement page and passes in the TA, Instructor and Admin, Profile and Button lists.

The determineRole method takes in a user object and returns a number based on the group of the user. This method retrieves the manager, ta and instructor groups, and compares them against the first group in the current User objects group list. It returns 1 for Admin, 2 for Instructor, 3 for TA and -1 for something going wrong, like a user without a group.

The load method takes in a user object, and determines the role of the user based on the return of the determinerole method. If the user is an admin, the method returns a list containing a list of every User object, a list of every profile object, and 2 lists of a create and a delete button respecively. If the user is an Instructor or a TA, the only non-empty list returned is the first list, which contains every User object.

I created a button model to represent the presence of buttons on a page, so that they could be tied to stuff in the backend.

I also created unit tests for the new helper methods, and updated our old acceptance tests to work with the new get method.